v1.3.8 formatted

+++ src/pointer/scroll.rs
@@ -1,5 +1,5 @@
-use std::collections::HashMap;
 use smithay_client_toolkit::seat::pointer::{PointerEvent, PointerEventKind};
+use std::collections::HashMap;
 use wayland_client::backend::ObjectId;
 
 use crate::AppState;
@@ -15,7 +15,7 @@ pub fn handle_scroll_events(
     let (dock_height, box_size, spacing, start_offset_x) = layout;
     let mut layer_changed = false;
 
-    let dock_top_bound = (state.height as i32).saturating_sub(dock_height);
+    let dock_top_bound = state.height.saturating_sub(dock_height);
     let is_over_icons = state.interaction.pointer_position.y >= dock_top_bound.into();
 
     let phys_width = (state.width as f32 * scale_factor).round() as i32;
@@ -36,7 +36,8 @@ pub fn handle_scroll_events(
 
                 if is_over_icons {
                     for (index, app_id) in apps_in_dock.iter().enumerate() {
-                        let start_x = start_offset_x + spacing + index as i32 * (box_size + spacing);
+                        let start_x =
+                            start_offset_x + spacing + index as i32 * (box_size + spacing);
                         let hit_start_x = start_x.saturating_sub(spacing / 2);
                         let hit_end_x = start_x + box_size + (spacing / 2);
 
@@ -53,22 +54,23 @@ pub fn handle_scroll_events(
                     if let Some(ref app_id) = state.hover_state.app_id {
                         if let Some(wins) = running_by_app.get(app_id) {
                             let total_apps = apps_in_dock.len();
-                            let hovered_app_index = apps_in_dock
-                                .iter()
-                                .position(|id| id == app_id)
-                                .unwrap_or(0);
+                            let hovered_app_index =
+                                apps_in_dock.iter().position(|id| id == app_id).unwrap_or(0);
 
-                            let (menu_x, menu_y, menu_width, menu_height, _, _) = geometry.compute_bounds(
-                                phys_width,
-                                phys_height,
-                                total_apps,
-                                hovered_app_index,
-                                wins.len(),
-                                scale_factor as f64,
-                            );
+                            let (menu_x, menu_y, menu_width, menu_height, _, _) = geometry
+                                .compute_bounds(
+                                    phys_width,
+                                    phys_height,
+                                    total_apps,
+                                    hovered_app_index,
+                                    wins.len(),
+                                    scale_factor as f64,
+                                );
 
-                            let ptr_x = (state.interaction.pointer_position.x as f32 * scale_factor).round() as i32;
-                            let ptr_y = (state.interaction.pointer_position.y as f32 * scale_factor).round() as i32;
+                            let ptr_x = (state.interaction.pointer_position.x as f32 * scale_factor)
+                                .round() as i32;
+                            let ptr_y = (state.interaction.pointer_position.y as f32 * scale_factor)
+                                .round() as i32;
 
                             if ptr_x >= menu_x
                                 && ptr_x <= menu_x + menu_width
