+++ src/render/text.rs
@@ -1,5 +1,6 @@
 use crate::FontManager;
 
+#[allow(clippy::too_many_arguments)]
 pub fn draw_text(
     canvas: &mut [u8],
     canvas_width: i32,
@@ -17,12 +18,15 @@ pub fn draw_text(
         let (metrics, bitmap) = font.rasterize(c, size as f32);
         for y in 0..metrics.height {
             for x in 0..metrics.width {
-                let canvas_x: i32 = (x_offset as i32 + x as i32 + metrics.xmin as i32) as i32;
-                let canvas_y: i32 = (start_y as i32 + y as i32 + (size as i32 - metrics.height as i32 - metrics.ymin as i32)) as i32;
+                let canvas_x: i32 = x_offset + x as i32 + metrics.xmin;
+                let canvas_y: i32 =
+                    start_y + y as i32 + (size - metrics.height as i32 - metrics.ymin);
 
-                if canvas_x < canvas_width as i32 && canvas_y < canvas_height as i32 {
-                    let canvas_idx: usize = ((canvas_y * canvas_width as i32 + canvas_x) * 4) as usize;
-                    let alpha = bitmap[(y as i32 * metrics.width as i32 + x as i32) as usize] as f32 / 255.0;
+                if canvas_x < canvas_width && canvas_y < canvas_height {
+                    let canvas_idx: usize = ((canvas_y * canvas_width + canvas_x) * 4) as usize;
+                    let alpha = bitmap[(y as i32 * metrics.width as i32 + x as i32) as usize]
+                        as f32
+                        / 255.0;
                     if alpha > 0.0 {
                         let b = color.2 as f32;
                         let g = color.1 as f32;
@@ -32,7 +36,7 @@ pub fn draw_text(
                         let cur_g = canvas[canvas_idx + 1] as f32;
                         let cur_r = canvas[canvas_idx + 2] as f32;
 
-                        canvas[canvas_idx]     = ((b * alpha) + (cur_b * (1.0 - alpha))) as u8;
+                        canvas[canvas_idx] = ((b * alpha) + (cur_b * (1.0 - alpha))) as u8;
                         canvas[canvas_idx + 1] = ((g * alpha) + (cur_g * (1.0 - alpha))) as u8;
                         canvas[canvas_idx + 2] = ((r * alpha) + (cur_r * (1.0 - alpha))) as u8;
                         canvas[canvas_idx + 3] = 255;
